保存代码
用这种方式是可行，softmax应该也可行，目前是3000个汉字

+++ ocr/ocr_data_collect.py
@@ -1,18 +1,21 @@
 #! /usr/bin/python3.12
 
-import os, config, openpyxl, shutil, random
+import os, config, openpyxl, shutil, random, threading
 from PIL import Image, ImageDraw, ImageFont,ImageFilter
 
 cur_dir = os.path.dirname(__file__)
 image_path = os.path.join(config.get_root_path(), 'ocr_datas', 'images')
 if os.path.exists(image_path):
-    print(f'begin clear {image_path}')
-    shutil.rmtree(image_path)
-    print(f'clear {image_path} success')
+    if input('目录已经在，是否清空目录?(Y)').strip() == 'Y':
+        print(f'begin clear {image_path}')
+        shutil.rmtree(image_path)
+        print(f'clear {image_path} success')
 os.makedirs(image_path, exist_ok=True)
 charset_file = os.path.join(config.get_root_path(), 'ocr_datas', 'charset.txt')
 
-samples_per_char = 10
+samples_per_char = 20
+samples_chars = None
+samples_thread_count = 20
 
 def load_fonts():
     font_path = os.path.join(cur_dir, 'fonts')
@@ -38,16 +41,72 @@ def load_ori_datas():
     print(f'sheet name {workbook.sheetnames}')
     # 把数据拆分成训练数据和验证数据
     charsets = []
+    chars = []
     for row in sheet.iter_rows(min_row=2, min_col=2, max_col=2):
         for cell in row:
             if cell.value is not None:
-                create_image(cell.value, charsets)
+                chars.append(cell.value)
+    chars = chars[:3000]
+    if samples_chars is None or samples_chars >= len(chars):
+        total = len(chars)
+        # 使用多线程生成
+        per_count = total // samples_thread_count
+        if total % samples_thread_count > 0:
+            per_count = per_count + 1
+        threads = []
+        for i in range(per_count):
+            sub_chars = chars[i * per_count : min((i + 1) * per_count, total)]
+            t = threading.Thread(target = thread_create_image, args=(sub_chars, i))
+            threads.append(t)
+        for t in threads:
+            t.start()
+        for t in threads:
+            t.join()
+        """created = 0
+        for ch in chars:
+            create_image(ch, charsets)
+            created = created + 1
+            print(f'create: ({created} / {total})')
+        """
+    else:
+        # 随机选择几个
+        count = 0
+        select_index = []
+        created = 0
+        total = len(samples_chars)
+        while(count < samples_chars):
+            index = random.randint(0, len(chars))
+            if index in select_index:
+                continue
+            create_image(chars[index], charsets)
+            count = count+ 1
+            print(f'create: ({created} / {total})')
+
 
     with open(charset_file, 'w+') as w_file:
         for ch in charsets:
             w_file.write(ch + '\n')
 
-def generate_char_image(ch, font):
+def image_change(image, size):
+    # 可选：轻度模糊 / 旋转
+    if random.random() < 0.5:
+        image = image.filter(ImageFilter.GaussianBlur(radius=0.2))
+
+    # 旋转
+    if random.random() < 0.5:
+        angle = random.randint(-30, 30)
+        image = image.rotate(angle, expand=True, fillcolor=255)
+
+    # 缩放
+    if random.random() < 0.2:
+        if random.random() < 0.5:
+            image = image.resize((int(size / 0.67), int(size / 0.67)), resample = Image.Resampling.LANCZOS)
+        else:
+            image = image.resize((size * 2, size * 2), resample = Image.Resampling.LANCZOS)
+
+    return image
+
+def generate_char_image(ch, font, transform=True):
     """
     根据文字创建图片，png
     :param ch 文字
@@ -65,25 +124,19 @@ def generate_char_image(ch, font):
     y = (size - h) / 2 + dy
     draw.text((x, y), ch, font=font, fill='black')
 
-    # 可选：轻度模糊 / 旋转
-    if random.random() < 0.5:
-        image = image.filter(ImageFilter.GaussianBlur(radius=0.5))
-
-    # 旋转
-    if random.random() < 0.5:
-        angle = random.randint(-30, 30)
-        image = image.rotate(angle, expand=True, fillcolor=255)
-
-    # 缩放
-    if random.random() < 0.3:
-        if random.random() < 0.5:
-            image = image.resize((int(size / 2), int(size / 2)), resample = Image.Resampling.LANCZOS)
-        else:
-            image = image.resize((size * 2, size * 2), resample = Image.Resampling.LANCZOS)
+    if transform:
+        image = image_change(image, size)
 
     return image
 
+def thread_create_image(chars, i):
+    charsets = []
 
+    for index in range(len(chars)):
+        create_image(chars[index], charsets)
+        print(f'thread-{i}, create image: {index+1} / {len(chars)}')
+
+    return charsets
 def create_image(ch, charsets):
     """
     根据文字创建图片，png
@@ -92,9 +145,11 @@ def create_image(ch, charsets):
     max_count = samples_per_char * len(fonts)
     for i in range(max_count):
         for font in fonts:
-            img = generate_char_image(ch, font)
             filename = f"{ch}_{i}.png"
             path = os.path.join(image_path, filename)
+            if os.path.exists(path):
+                continue
+            img = generate_char_image(ch, font)
             img.save(path)
             charsets.append(f"{filename}\t{ch}")
